Refactor Scene into modular components
